core: move TestContractCall into the contract package
It belongs there and now it can be moved. No functional changes.
